9683 Allow bypassing devid in vdev_disk_open()
Reviewed by: Sara Hartse <sara.hartse@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

We've had 2 cases where the devids of disks were shuffled, preventing
domain0 from being imported, which required a tedious manual process to
zero-out devids through kmdb during zpool import.

This feature would add a tunable to vdev_disk_open() to bypass devids
and add more debug messages.  Note that if we bypass the devid by
setting vdev_disk_bypass_devid, the pool will not have a devid field in
the config anymore. Even if we re-import the pool with
vdev_disk_bypass_devid unset, the devid field will never be written to
the config again, so once it's gone it is gone for good.