9691 fat zap should prefetch when iterating
Reviewed by: Serapheim Dimitropoulos <serapheim.dimitro@delphix.com>
Reviewed by: Pavel Zakharov <pavel.zakharov@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

When iterating over a ZAP object, we're almost always certain to iterate over
the entire object.  If there are multiple leaf blocks, we can realize a
performance win by issuing reads for all the leaf blocks in parallel when the
iteration begins.

For example, if we have 10,000 snapshots, "zfs destroy -nv pool/fs@1%9999" can
take 30 minutes when the cache is cold. This change provides a >3x performance
improvement, by issuing the reads for all ~64 blocks of each ZAP object in
parallel.