fix a bunch of formatting issues
Some were introduced in this branch while others existed
in the changed files before the branch.
Most notably, comments that describe options of zpool commands
now should use tabs for formatting.  Previously some option
descriptions were formatted with tabs, some with spaces and some
with a mix of tabs and spaces.

Finally, the branch is rebased on the latest master.