9440 Decrease contention on dn_struct_rwlock
Reviewed by: Matt Ahrens <matt@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

We created a new rwlock in each dbuf that we use to protect its contents
if it contains metadata. This allows significantly increased parallelismxi
across locks while maintaining safety.

Then we fixed a problem in lock ordering in the initial commit. The fix is
to acquire the child's mutex before acquiring the parent's rwlock in all
cases.

Upstream bug: DLPX-52564 DLPX-53085